Refactor: `roles` resolver latest subgraph block
