+++ src/data/graphql/typeDefs/subgraphGeneral.ts
@@ -1,6 +1,16 @@
 import gql from 'graphql-tag';
 
 export default gql`
+  type SubgraphBlock {
+    id: String!
+    timestamp: String!
+  }
+
+  type SubgraphTransaction {
+    id: String!
+    block: SubgraphBlock!
+  }
+
   extend type Query {
     events(
       skip: Int
@@ -9,5 +19,6 @@ export default gql`
       orderDirection: String
       block: ToBlockInput
     ): [SubgraphEvent!]!
+    block(id: String!): SubgraphBlock
   }
 `;
